refactor: modularize GUI components
